commit 07-12-23 pt1

+++ models/budgeting/Analisa_Budgeting.go
@@ -6,8 +6,8 @@ type Analisa_Budgeting struct {
 	Biaya_Mingguan           int64                      `json:"biaya_mingguan"`
 	CV                       int64                      `json:"cv"`
 	SV                       int64                      `json:"sv"`
-	CPI                      int64                      `json:"cpi"`
-	SPI                      int64                      `json:"spi"`
+	CPI                      float64                    `json:"cpi"`
+	SPI                      float64                    `json:"spi"`
 	Detail_Analisa_Budgeting []Detail_Analisa_Budgeting `json:"detail_analisa_budgeting"`
 }
